chore: Update Actions workflow to use external actions.
This change refactors the main workflow to use VaultVulp/action-pipenv
to execute Pipenv commands. Additionally, support for the Windows
Pipenv scripts (although not the matrix entries to run them) is
implemented. Finally, the main workflow now runs on all pushs and PRs.

In addition to the included Windows changes, VaultVulp/action-pipenv
does not support Windows runners, and so an alternative will be needed.